Update the tests to account for the fact that oauth headers are between double quotes
The code that extracts the value of the oauth parameter is updated to validate the presence of the double quotes and strip them before returing the value of the header.
With this change, tests don't have to validate the parameter is between double quotes and instead can validate the actual value directly.